A lot of changes
List of changes:
- SecureRandom fixes ("PRNGFixes") applied for Android 4.3 and below.
- App is now able to force runtime permissions and authorizations (like
  Device Admin or System Overlays) as long as root permissions are given
  to it - THOUGH the feature is disabled for now (there must be a way to
  request VISOR to stop forcing them if the user so chooses).
  PS: currently forcing Device Admin is only available for Lollipop+.
- Device Admin is attempted to be forced every time the application or
  the Main Service start.
- The project Java version restriction is now removed. I'm not too sure
  why I put it here in the first place. Confusion with something else,
  probably.
- Battery state (charging/not charging) and battery presence improved
  to support devices that don't broadcast certain broadcasts, or that
  don't carry some correct values on them.
- PocketSphinx recognition files updated. This is the reason for the app
  to need 100 MB now... I've added good and low quality files and kept
  the already present ones too.
- Improved the privileged app installer script to auto detect the device
  characteristics and install the app as VISOR on the system folders.
- ModulesList class vastly improved. Module support checks are now in
  place automatically before starting a module, proper module check has
  been added (now each module implements a function that must return if
  it's supported or not, whatever that means for that specific module -
  can include permissions check, hardware check, external apps check,
  whatever that is needed for it to work --at a minimum level--,
  because the CameraManagement will be able to record videos, but
  videos don't necessarily need audio too).
- Development permissions have now been added to the app. They won't
  be granted without ADB commands or through automatic forcing with root
  permissions though (which are disabled anyways for now, so only ADB).
- Since the app now properly checks if a module is supported or not with
  permissions (and the rest), permissions checks are now in place too to
  check if they have been granted to the app (previously, I just assumed
  they were given ("God mode")).
- Protected Lock Screen improved a bit. Fixed possible PLS auto-restarts
  after dismissing it.
- Full control of the speech recognizers given to SpeechRecognitionCtrl,
  as should have been done already. Previously I had it separated in
  case the controller would crash or something - that's a mess, no way.
  Only the controller can control the recognizers now (except
  terminating them, which must be done at the instant of the function
  call, currently).
- Android utilities classes improved.
- All file and directory operations have been changed to be based on
  shell commands only. Might change that later. Currently each call
  starts a new shell process. That's a bit too much sometimes. Either
  all calls are redirected to one global shell process or this class is
  changed to only use shell commands in case there is no permission
  using the SDK methods.
- Java reflection utilities class added to ease reflection usage.
- Root check changed. Now only "has" or "doesn't have" root permissions.
- Shell commands executor function improved. Might be faster now, and in
  any case, it's more organized.

+++ app/src/main/java/com/dadi590/assist_c_a/GlobalUtils/UtilsProcesses.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 DADi590
+ * Copyright 2022 DADi590
  *
  * Licensed to the Apache Software Foundation (ASF) under one
  * or more contributor license agreements.  See the NOTICE file
@@ -57,11 +57,14 @@ public final class UtilsProcesses {
 
     /**
      * <p>Terminates a PID.</p>
+     * <p>If -1 is provided, nothing is done.</p>
      *
      * @param pid PID to terminate
      */
     public static void terminatePID(final int pid) {
-        android.os.Process.killProcess(pid);
+        if (-1 != pid) {
+            android.os.Process.killProcess(pid);
+        }
     }
 
     /**
